modificacion en las caracteristicas de las rutas de usuarios

+++ controller/users.controller.js
@@ -53,20 +53,25 @@ module.exports = {
       pageSize: limit,
       numberOfPages: numberOfPages,
     };
+    res.append('pagination', response.pagination)
 
     if (startI > 0) {
-      response.link = {
-        first: `/users?_page=1&_limit=${limit}`,
-        prev: `/users?_page=${page - 1}&_limit=${limit}`,
-      };
+      // response.link = {
+      //   first: `/users?_page=1&_limit=${limit}`,
+      //   prev: `/users?_page=${page - 1}&_limit=${limit}`,
+      // };
+      res.append('fist', `/users?_page=1&_limit=${limit}`)
+      res.append('prev', `/users?_page=${page - 1}&_limit=${limit}`)
     }
 
     if (endI < users.length) {
-      response.link = {
-        ...response.link,
-        next: `/users?_page=${page + 1}&_limit=${limit}`,
-        last: `/users?_page=${numberOfPages}&_limit=${limit}`,
-      };
+      // response.link = {
+      //   ...response.link,
+      //   next: `/users?_page=${page + 1}&_limit=${limit}`,
+      //   last: `/users?_page=${numberOfPages}&_limit=${limit}`,
+      // };
+      res.append('next', `/users?_page=${page + 1}&_limit=${limit}`)
+      res.append('last', `/users?_page=${numberOfPages}&_limit=${limit}`)
     }
     response.result = users.slice(startI, endI).map((user) => ({
       _id: user._id,
@@ -74,7 +79,7 @@ module.exports = {
       role: user.role,
     }));
 
-    return res.json(response); 
+    return res.json(response.result); 
 
 
     } catch (err) {
@@ -101,12 +106,17 @@ module.exports = {
     const user = idOrEmail(req.params.uid);
     const userUp = {
       email: req.body.email,
-      password : user.password,
+      password : req.body.password,
       role: req.body.role,
     }
     
-
     try {
+      
+      if (!userUp.email || !userUp.password ) return next(400);
+
+      const userFound = await User.findOne({ email: user.email });
+      if (!userFound) return next(404);
+
       userUp.password = bcrypt.hashSync(req.body.password, 10);
       
       const newUsr = await User.findOneAndUpdate(user, userUp, {
@@ -120,7 +130,7 @@ module.exports = {
       });
 
     } catch (error) {
-      return next(404);
+      return res.json({ error: error.message });
     }
 
   },
@@ -128,6 +138,9 @@ module.exports = {
     const user = idOrEmail(req.params.uid);
     
     try {
+      const userFound = await User.findOne( user);
+      if (!userFound) return next(404);
+
       const deletedUser =await User.findOneAndDelete(user).lean();
 
       return res.json({
@@ -137,7 +150,7 @@ module.exports = {
       });
    }
    catch(e){
-    return next(404);
+    return res.json({ error: error.message });
    }
   }
 };
